fix(end): Implemented droop tally
There was a case where the number of seats was significantly higher than the number of options in a question, which resulted in fewer seats than desired. This has been fixed so that, if it exceeds the number of seats, there are options that can be increased by more than 1, for which a Counter is used.