feat(init): Integration of score voting
Score voting was already implemented by other developer but this one left the working group.
As a consecuence, the feature was not updated and needed to be integrated with other types of voting.
It has been moved the funcionality of the algorithm from model to view ir order to follow the standard.
It has been updated the attribute name according to a standard format.